Refactoring index.tsx and Modal.tsx. Created a new component AcceptDeclineModal.tsx
